xdg: Add support for basedir spec
It is very common (in Linux) to store files and folders according to the
Basedir spec. The pyxdg module handles all the hard parts, and the
options variables kind-of already existed, so lets put them to better
use.

This only works properly for the main application, as some other
packages also include their own main.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>